+++ sources/Google.Solutions.IapDesktop.Extensions.Shell.Test/Google.Solutions.IapDesktop.Extensions.Shell.Test.csproj
@@ -168,8 +168,8 @@
     <Compile Include="Views\SshTerminal\TestSshTerminalView.cs" />
     <Compile Include="Views\SshTerminal\TestSshTerminalSessionBroker.cs" />
     <Compile Include="Views\SshTerminal\TestSshTerminalPaneViewModel.cs" />
-    <Compile Include="Views\TestConnectCommands.cs" />
-    <Compile Include="Views\TestSessionCommands.cs" />
+    <Compile Include="Views\Session\TestConnectCommands.cs" />
+    <Compile Include="Views\Session\TestSessionCommands.cs" />
     <Compile Include="Views\TestSessionPaneBase.cs" />
     <Compile Include="Views\TunnelsViewer\TestTunnelsViewCommands.cs" />
     <Compile Include="Views\TunnelsViewer\TestTunnelsViewModel.cs" />
